fixup! Bugfix: Syncthing would not shutdown on exit
https://github.com/sieren/QSyncthingTray/issues/66#issuecomment-174223213

+++ src/syncconnector.cpp
@@ -486,8 +486,12 @@ void SyncConnector::killProcesses()
   if (mpSyncProcess != nullptr
       && mpSyncProcess->state() == QProcess::Running)
   {
-    mpSyncProcess->kill();
-    mpSyncProcess->waitForFinished();
+    shutdownSyncthingProcess();
+    bool hasFinished = mpSyncProcess->waitForFinished(10000);
+    if (!hasFinished)
+    {
+        mpSyncProcess->kill();
+    }
   }
   if (mpSyncthingNotifierProcess != nullptr
       && mpSyncthingNotifierProcess->state() == QProcess::Running)
